Add parse button to flashlog for parsing
